Address Andre's review comments
- set thread affinity on the `ping_thread` according to the config
- move network utility functions to `lib/utils/networkFunctions`
- follow Doxygen standard in documenting public methods
